+++ src/hull/mod.rs
@@ -1,0 +1,3 @@
+mod convex_hull2;
+
+pub use convex_hull2::{HullBoundaryPoints, convex_hull2};